fix: fixed tabs content rendering
