+++ my-first-app/src/Components/Tabs/TabContent/TabContent.module.scss
@@ -29,54 +29,9 @@
     }
   }
 }
-.NumberButton {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-  font-size: 16px;
-  line-height: 24px;
-  font-weight: 600;
-  color: var(--text-color);
-  gap: 8px;
-}
 
 .icon {
   width: 18px;
   height: 14px;
   fill: var(--text-color);
-}
-
-.pageNubmers {
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-  font-size: 16px;
-  line-height: 24px;
-  font-weight: 600;
-  color: var(--text-color);
-}
-.pageNubmersCurr {
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-  font-size: 16px;
-  line-height: 24px;
-  font-weight: 600;
-  color: #2231aa;
-}
-
-.pageNubmersWrap {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 32px;
-}
-
-.numbersWrapper {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
 }
