+++ my-first-app/src/Components/Tabs/Tabs.tsx
@@ -14,7 +14,7 @@ const Tabs = () => {
       <div className={style.tabListWrap}>
         <Tablist selectedTab={showTab} setShowTab={handler} />
       </div>
-      <TabContent />
+      <TabContent selectedTab={showTab} />
     </>
   );
 };
